[FINNAPOSTI-276] Finding Primo Central OpenURLs
Apparently OpenURLs are not always included in the sear:LINKS section
of Primo Central records, and we need to look for them in the GetIt2
and GetIt1 attributes of the sear:GETIT element. But at least GetIt1
can also include direct, non-OpenURL links, so a check was added that
we actually got an OpenURL 1.0 link.